Mummi and bilbo contribution!! veriverse contribution!! LCTV contribution!! catherinacortega, sjpablo (Pawel Kulesza), fuel (Benjamin Kuhl), r00tkiller (Alexandre),  vapaamies (Vladislav Javadov), epryka (Vova - Vladimir), theaifam5 (Mateus), doraemon (currently... also gosu - hiiamgosu, dum - noticemesenpaiiamdum..., litAsHellhahah), mushmouth (Patrick McFarling),  sonnystar (sonnystaminecraft, Vlad Ionescu), ThomasBS, abbyyyyy, alekcy, alexcode, anita, ayvatan, bazz3l, bigboom, blackburn1911, blackhatisme, boza, camwhite (Cameron White - YOSHICREW, speakto.space | ng2/node + mixing fire, Yoshicrew.com-out-loud contribution...), chrifrachid, codeotea, collins_badmus, cookiesliyr, chubz, csharper, cubor, danielv, david_joffe, davids81, denvercowboy, dgazail, diagram, dopedopedope, drmjg, dzetko, dzhekson6000, EckyYakov, euphemisminc, fauker, ffabiorj, flythiago, georgeb, gillie, gzj71ck4elm2, hailviral1, hasanakyildiz, heuy_freeman, hudwan, igsky, intelijens (Michael Bui), ishaan007lucky, jarzi, jaycee, jilizart, johnyblaze, jookeer, kaeris, karenkododyan, karrmarr, kaznowski, kissloryshy, kranalex, LaffayetteBrittoPhoto (streamer with its C300 photo), livecodingtv..., livetester, magrib, marcodieba, marianyc, matroskinb, metalplay, michaeljcalkins, michaelphp, monicaa,  nuke, ollla (Olya Medeveva), paezrice, pavankataria, peter_cados, phanxgames (Henry Price), phreak10, pulkas, purpleizlegitx, rex2, rheghal, rooney2, sanyakozmey, senhorbardell, sharpadogge, silvaesouza, simplyronan, siox, skyfoxx, smart6307 (thunder will/canal2048), sova, steffen25, systemis, szy, tarikogut, thiagovgamg, tinyvane, uberram, usandy, v0n3x, wanket, wapaca, web2ls, yanaseattle, yurihoyos, zaaz, zakaria98 and zzzel (Marko Zel)  contribution!! jiangyubo530, kozmo, contribution!! Twitch contribution!! AlfredoPlays (and his mates), Tectuma (GothicatCat1(Stacie), Tectuma(Kevin), Kenny, Martin, Matty and Dylan), Above_Perfection, Bialyorzeu, cary205, cpmgamingltd, Curseappbot, DuckofDoom12, GameDevCompany (Björn), GrubbyJuicde, hailviral1, Handmadehero (Casey Muratori), jhovgaard, jordanpop33, Letorusii, Mattyplays6, michaeljcalkins, nothings2 (Sean Barrett), Shathan, Skintrader_pl, Snipe0803, Sweegzv, themushmouth and XanderGamingTwich contribution!! Beam.pro contribution!! danielfdeus, RoxenRox and Xellar contribution!! bisba contribution!! Carol, Lúcio and Lucinho contribution!! Santos contribution!! Paulo Picciani contribution!!
Thank you Camilo!

bilbo at veriverse
i7 6700 32GB RAM Windows 10 Pro
(booting also macOS Sierra, Hyper-V and Ubuntu 16.04 Unity Numix Theme Numix Circle Icons)

+++ appseed/axis/axis/node/windows/windows_ip_enum.cpp
@@ -217,7 +217,7 @@ namespace windows
                
                {
                   auto ipv6 = (struct sockaddr_in6 *) ptr->ai_addr;
-                  ipa.add(::net::address(*ipv6,ptr->ai_addrlen));
+                  ipa.add(::net::address(*ipv6,(int)ptr->ai_addrlen));
                }
